+++ __tests__app__/app/absolute-test/page.tsx
@@ -1,0 +1,3 @@
+import { Page } from '../../../src/app';
+
+export default class PagePage extends Page {}