+++ __tests__app__/dev/src/plugin.js
@@ -1,8 +1,0 @@
-import { Plugin } from '../../../server';
-
-export default class ImportedPlugin extends Plugin {
-
-	getName() {
-		return 'Imported plugin';
-	}
-}